backend/requirements.txt
fastapi
uvicorn
torch
transformers
python-dotenv
groq